backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
email-validator
pymongo
motor
python-jose
cloudinary
python-dotenv
